+++ plone/app/content/browser/contents/__init__.py
@@ -90,9 +90,10 @@ class ContentsBaseAction(BrowserView):
         context = aq_inner(self.context)
         selection = self.get_selection()
 
-        parts = str(self.request.form['folder'].lstrip('/')).split('/')
-        parent = self.site.unrestrictedTraverse('/'.join(parts[:-1]))
-        self.dest = parent.restrictedTraverse(parts[-1])
+        parts = str(self.request.form.get('folder', '').lstrip('/')).split('/')
+        if parts:
+            parent = self.site.unrestrictedTraverse('/'.join(parts[:-1]))
+            self.dest = parent.restrictedTraverse(parts[-1])
 
         self.catalog = getToolByName(context, 'portal_catalog')
         self.mtool = getToolByName(self.context, 'portal_membership')